python/stats/requirements.txt
grpcio==1.10.0
pika==0.11.2
protobuf==3.5.2